Work around known spec bug for MouseEvent.offsetX/Y
Known spec bug (WD 3 August 2016):
The behavior in the spec doesn't do "during and after the dispatch".
It just does "during the dispatch", and goes back to "pageX" after the
dispatch. This is not expected (and cf. implemented) behavior, so we
don't follow the spec here to align the property with all other browsers.

See also:
https://www.w3.org/Bugs/Public/show_bug.cgi?id=16673